requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-ckeditor==6.7.0
django-import-export==3.3.1
django-js-asset==2.1.0
django-map-widgets==0.4.2
django-tastypie==0.14.6
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
xlrd==2.0.1
xlwt==1.3.0
